requirements.txt
openai>=1.0.0
numpy>=2.0.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
